Major Optimizations
* Now checks thermostatSummary to see if anything has changed
* Only gets changed runtime and/or thermostat objects if/when they change (independently) - significantly reduces heavy API calls for data that doesn't change frequently (settings, programs, events)
* Each thermostat object is stored (and accessed) separately so that they can be independently updated
* Sends equipmentStatus and humiditySetpoint to thermostat device(s)
* Eliminated most time checks, opting instead to let thermostatSummary govern if/when we poll for thermostat data

NOTE: A minimum poll time of 3 minutes is still wise, but this has also been tested with poll = 1 minute. Watchdog devices can also be used safely now, thanks to thermostatSummary.

SUGGESTION: Add a few of your indoor SmartThings devices that report temperature or humidity to the watchdog list to reduce the time between when the HVAC stops/starts and the SmartDevices report the changes.