Major new optimization
Notonly updates only the thermostat/runtime objects that have changed, but now will only request those updates for the thermostat(s) that have changed.

Originally, this code would get ALL the data for ALL the thermostats every poll cycle, now we are down to only getting data changes for individual thermostats. This SIGNIFICANTLY reduces not only the traffic to ecobee, but also the traffic between the SmartApp and the device itself.